+++ packages/client/verifiable-credentials/src/presentation/getCredentialNfts.ts
@@ -1,11 +1,7 @@
-import {
-    CredentialFilter,
-    VCChain,
-    getUsersCredentialNfts as getUsersCredentialNftsRaw,
-} from "@/verifiableCredentialsSDK";
+import { CredentialFilter, VCChain, getCredentialNfts as getCredentialNftsRaw } from "@/verifiableCredentialsSDK";
 
 import { getWalletVcCompatibleNfts } from "./getNfts";
 
-export async function getUsersCredentialNfts(chain: VCChain, wallet: string, filters: CredentialFilter = {}) {
-    return getUsersCredentialNftsRaw(chain, wallet, getWalletVcCompatibleNfts, filters);
+export async function getCredentialNfts(chain: VCChain, wallet: string, filters: CredentialFilter = {}) {
+    return getCredentialNftsRaw(chain, wallet, getWalletVcCompatibleNfts, filters);
 }
