+++ packages/client/verifiable-credentials/src/verifiableCredentialsSDK/index.ts
@@ -9,8 +9,8 @@ export { verifyCredential } from "./verification";
 // Presentation
 export { CredentialService, ipfsRetrievalProcedure } from "./presentation/getCredential";
 export { ContractMetadataService } from "./presentation/contractMetadata";
-export { getUsersCredentialNfts } from "./presentation/getCollections";
-export { getNFTFromLocator } from "./presentation/nftByLocator";
+export { getCredentialNfts } from "./presentation/getCollections";
+export { getCredentialNFTFromLocator } from "./presentation/nftByLocator";
 
 // Decryption
 export { Lit, LitNetwork } from "./encryption/lit";
